Updated Makefile and added Shulker version
The Makefile has been updated to include a new target for installing Shulker. A new file was also created to keep track of the current Shulker version. The changes ensure that the Shulker installation is part of the build process, and it's always using the correct version as specified in the .versions/shulker file.